Renaming config files to match original versions
