+++ lib/default-config.js
