+++ lib/get-config.js
@@ -1,6 +1,6 @@
 "use strict";
 
-var defaultConfig = require("./default_config");
+var defaultConfig = require("./default-config");
 
 module.exports = function getConfig(custom) {
     var config = {};
