+++ lib/test.js
@@ -9,7 +9,7 @@
 "use strict";
 
 var utils = require("./utils");
-var getConfig = require("./config");
+var getConfig = require("./get-config");
 var slice = Array.prototype.slice;
 
 function finish(sandbox, error, dontThrow) {
